kobrasuitecore/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
distlib==0.3.9
Django==5.1.4
django-redis==5.4.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.16.1
gunicorn==23.0.0
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.2.1
setuptools==75.6.0
sqlparse==0.5.3
virtualenv==20.28.0
whitenoise==6.6.0
